BUG FIX: Fixed the 'print_token' bug in Parser where the tokens needed to be cast to integers so they do not go invalid!

+++ src/parser.c
@@ -76,7 +76,7 @@ void print_token(Token t) {
   if (t.end - t.beginning < 1) {
     printf("INVALID TOKEN POINTERS");
   } else {
-    printf("%.*s", t.end - t.beginning, t.beginning);
+    printf("%.*s", (int)(t.end - t.beginning), t.beginning);
   }
 }
